Don't die on valid emails, but do exit with code 1
Previously, it would exit as soon as it found any bad email, and it
wouldn't send anything else, because it had errorred out. Catching
StandardError instead of RuntimeError fixed that, but then we weren't
getting emails when there were problems because it was exiting 0 after
logging the failures.

So I added the ugly lines tracking whether we've failed to send any
emails because I wasn't willing to try and refactor this.